+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 
 - [home] feature: new `/home` root command, to teleport to any of the player's homes.
 - [command_rewrite] feature: pre-compile the `regex` for better performance.
+- [cleaner] fix: a possible deadlock while iterating the entities in a world.